+++ src/main/java/io/github/forgestove/create_cyber_goggles/core/event/ItemTooltip.java
@@ -1,0 +1,26 @@
+package io.github.forgestove.create_cyber_goggles.core.event;
+import com.simibubi.create.content.equipment.armor.*;
+import com.simibubi.create.foundation.utility.CreateLang;
+import io.github.forgestove.create_cyber_goggles.core.util.CCGLang;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.event.entity.player.ItemTooltipEvent;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+public class ItemTooltip {
+	public static void tick(@NotNull ItemTooltipEvent event) {
+		var stack = event.getItemStack();
+		var tooltip = event.getToolTip();
+		backtank(stack, tooltip);
+	}
+	private static void backtank(@NotNull ItemStack stack, List<Component> tooltip) {
+		if (!(stack.getItem() instanceof BacktankItem)) return;
+		var component = CreateLang.translate("gui.goggles.fluid_container.capacity")
+			.style(ChatFormatting.GRAY)
+			.add(CCGLang.fraction(BacktankUtil.getAir(stack), BacktankUtil.maxAir(stack)))
+			.component();
+		tooltip.add(1, component);
+	}
+}