+++ src/main/java/io/github/forgestove/create_cyber_goggles/core/event/KeyInput.java
@@ -4,7 +4,7 @@ import com.simibubi.create.*;
 import com.simibubi.create.content.logistics.filter.*;
 import com.simibubi.create.content.logistics.stockTicker.*;
 import io.github.forgestove.create_cyber_goggles.*;
-import io.github.forgestove.create_cyber_goggles.core.util.CCGUtil;
+import io.github.forgestove.create_cyber_goggles.core.util.*;
 import me.shedaniel.autoconfig.AutoConfig;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;